+++ agent/loading-bay/metastore/core/src/main/java/com/hotels/road/hive/metastore/LocationResolver.java
@@ -18,5 +18,5 @@ package com.hotels.road.hive.metastore;
 import java.net.URI;
 
 public interface LocationResolver {
-  URI resolveLocation(String location);
+  URI resolveLocation(String location, boolean create);
 }