+++ agent/loading-bay/metastore/s3/src/main/java/com/hotels/road/hive/metastore/s3/S3LocationResolver.java
@@ -17,33 +17,45 @@ package com.hotels.road.hive.metastore.s3;
 
 import static java.util.stream.Collectors.joining;
 
+import java.io.ByteArrayInputStream;
 import java.net.URI;
 import java.net.URISyntaxException;
 import java.util.ArrayList;
 import java.util.List;
 
-import org.apache.commons.lang.StringUtils;
-
 import lombok.RequiredArgsConstructor;
 
+import com.amazonaws.services.s3.AmazonS3;
+import com.amazonaws.services.s3.model.ObjectMetadata;
 import com.hotels.road.hive.metastore.LocationResolver;
 import com.hotels.road.hive.metastore.MetaStoreException;
 
+import org.apache.commons.lang.StringUtils;
+
 @RequiredArgsConstructor
 public class S3LocationResolver implements LocationResolver {
+  private final AmazonS3 s3;
   private final String bucket;
   private final String prefix;
 
   @Override
-  public URI resolveLocation(String location) {
+  public URI resolveLocation(String location, boolean create) {
     List<String> elements = new ArrayList<>();
     if (StringUtils.isNotBlank(prefix)) {
       elements.add(prefix);
     }
     elements.add(location);
-    String path = elements.stream().collect(joining("/", "/", ""));
+    String path = elements.stream().collect(joining("/"));
+
+    if (create) {
+      ByteArrayInputStream content = new ByteArrayInputStream(new byte[0]);
+      ObjectMetadata metadata = new ObjectMetadata();
+      metadata.setContentLength(0);
+      s3.putObject(bucket, path + "/", content, metadata);
+    }
+
     try {
-      return new URI("s3", bucket, path, null);
+      return new URI("s3", bucket, "/" + path, null);
     } catch (URISyntaxException e) {
       throw new MetaStoreException(
           String.format("Could not construct location URI for: bucket='%s', prefix='%s', location='%s'", bucket, prefix,